gpu: move the offload pre-flight into its own unit
The questions the pass asks before it commits to a kernel -- can the
backend size every per-thread workspace, can the device represent this
derived type or this scalar width, does a BLOCK-local array have an
extent anywhere, does the body run a statement the device has no way to
run, would a device callee need a run-time sized array temporary -- form
one group with one purpose: deciding whether to decline. They ask nothing
of the visitor and are asked from it.

Pure code motion: the definitions are unchanged, only their home is.
