+++ src/libasr/CMakeLists.txt
@@ -84,6 +84,7 @@ set(SRC
     pass/gpu_decline.cpp
     pass/gpu_offload.cpp
     pass/gpu_offload_designator.cpp
+    pass/gpu_offload_preflight.cpp
     pass/device_partition.cpp
     pass/device_launch_expand.cpp
     pass/gpu_memory_space.cpp
