+++ src/libasr/pass/gpu_offload.cpp
@@ -15,6 +15,7 @@
 #include <libasr/pass/device_launch_expand.h>
 #include <libasr/pass/gpu_decline.h>
 #include <libasr/pass/gpu_offload_designator.h>
+#include <libasr/pass/gpu_offload_preflight.h>
 #include <libasr/pass/intrinsic_array_function_registry.h>
 #include <libasr/pass/stmt_walk_visitor.h>
 #include <libasr/pass/pass_utils.h>
@@ -138,334 +139,6 @@ private:
 } // anonymous namespace
 
 static int gpu_kernel_counter = 0;
-
-// The scalar kernel parameter carrying one extent of an array parameter.
-// The kernel extraction creates it and the workspace resolver looks it up
-// again by name, so both spell it here.
-// The pieces of `size(s%m, d)` and `size(a(i)%m, d)`: the extent of one
-// dimension of an allocatable array component of a derived type, reached
-// either directly from a kernel argument or through a subscript into an
-// array of them.  Neither side of a kernel launch can reproduce such an
-// extent as it stands -- the kernel is handed the component as one flat
-// data buffer plus a per-element total size, never the component's
-// individual extents -- so the GPU offload pass turns it into a scalar
-// kernel argument instead.  The pass and the workspace resolver
-// recognise the shape through this one function, so what the pass
-// rewrites and what the resolver accepts cannot drift apart.
-struct GpuStructArrayMemberExtentRef {
-    // The subscript into an array of derived types, or nullptr when the
-    // component hangs off the kernel argument directly.
-    ASR::ArrayItem_t *item = nullptr;
-    // The kernel argument the component path is rooted at.
-    ASR::Var_t *base = nullptr;
-    ASR::StructInstanceMember_t *member = nullptr;
-    int64_t dim = 0; // 1-based
-};
-
-// The same shape, given the component reference and the 1-based
-// dimension on their own rather than wrapped in an `ArraySize`.  A
-// dimension of an array-valued intrinsic's operand is asked for this way:
-// the extent is derived from the operand, and no `size(...)` naming it
-// exists anywhere for the caller to hand over.
-static bool match_gpu_struct_array_member_extent_base(ASR::expr_t *expr,
-        int64_t d, const std::vector<std::string> &arg_names,
-        GpuStructArrayMemberExtentRef &out) {
-    if (d < 1) return false;
-    ASR::expr_t *base = ASRUtils::get_past_array_physical_cast(expr);
-    if (!base || !ASR::is_a<ASR::StructInstanceMember_t>(*base)) {
-        return false;
-    }
-    ASR::StructInstanceMember_t *sim =
-        ASR::down_cast<ASR::StructInstanceMember_t>(base);
-    ASR::expr_t *inner = ASRUtils::get_past_array_physical_cast(sim->m_v);
-    if (!inner) return false;
-    ASR::ArrayItem_t *ai = nullptr;
-    ASR::expr_t *root = inner;
-    if (ASR::is_a<ASR::ArrayItem_t>(*inner)) {
-        ai = ASR::down_cast<ASR::ArrayItem_t>(inner);
-        root = ASRUtils::get_past_array_physical_cast(ai->m_v);
-    }
-    if (!root || !ASR::is_a<ASR::Var_t>(*root)) return false;
-    std::string arr_name = ASRUtils::symbol_name(
-        ASR::down_cast<ASR::Var_t>(root)->m_v);
-    bool is_arg = false;
-    for (auto &a : arg_names) {
-        if (a == arr_name) { is_arg = true; break; }
-    }
-    if (!is_arg) return false;
-    ASR::ttype_t *mem_t =
-        ASRUtils::type_get_past_allocatable_pointer(sim->m_type);
-    if (!ASR::is_a<ASR::Array_t>(*mem_t)) return false;
-    if ((size_t)d > ASR::down_cast<ASR::Array_t>(mem_t)->n_dims) {
-        return false;
-    }
-    // Every subscript has to be reproducible in the host scope too: a
-    // compile-time integer constant (a named constant included), or a
-    // scalar the kernel is handed as an argument.
-    for (size_t k = 0; ai != nullptr && k < ai->n_args; k++) {
-        ASR::array_index_t &ix = ai->m_args[k];
-        if (ix.m_left || ix.m_step || !ix.m_right) return false;
-        int64_t sub_val;
-        if (try_eval_int_constant(ix.m_right, sub_val)) continue;
-        if (!ASR::is_a<ASR::Var_t>(*ix.m_right)) return false;
-        std::string sub_name = ASRUtils::symbol_name(
-            ASR::down_cast<ASR::Var_t>(ix.m_right)->m_v);
-        bool sub_is_arg = false;
-        for (auto &a : arg_names) {
-            if (a == sub_name) { sub_is_arg = true; break; }
-        }
-        if (!sub_is_arg) return false;
-    }
-    out.item = ai;
-    out.base = ASR::down_cast<ASR::Var_t>(root);
-    out.member = sim;
-    out.dim = d;
-    return true;
-}
-
-static bool match_gpu_struct_array_member_extent(ASR::expr_t *expr,
-        const std::vector<std::string> &arg_names,
-        GpuStructArrayMemberExtentRef &out) {
-    if (!expr || !ASR::is_a<ASR::ArraySize_t>(*expr)) return false;
-    ASR::ArraySize_t *as = ASR::down_cast<ASR::ArraySize_t>(expr);
-    if (!as->m_dim || !ASR::is_a<ASR::IntegerConstant_t>(*as->m_dim)) {
-        return false;
-    }
-    if (!ASR::is_a<ASR::Integer_t>(*ASRUtils::extract_type(as->m_type))) {
-        return false;
-    }
-    return match_gpu_struct_array_member_extent_base(as->m_v,
-        ASR::down_cast<ASR::IntegerConstant_t>(as->m_dim)->m_n,
-        arg_names, out);
-}
-
-// The value of the first whole-array assignment to `var_name` in `body`,
-// or nullptr.  A local that carries no shape anywhere else is sized by
-// what is assigned to it.
-// An array-valued operand of the elementwise expression `e`, or nullptr
-// when `e` is not built elementwise or has no array operand of its own.
-//
-// `sqrt((x-x0)**2 + (y(j)-y0)**2)` carries no dimensions in its type --
-// an elementwise result is a descriptor with empty extents -- but every
-// array operand of an elementwise expression is conformable with the
-// result, so any one of them has the result's shape.  A broadcast operand
-// is skipped: it is a scalar given a shape borrowed from elsewhere, so it
-// is never the operand that fixes the shape.
-static ASR::expr_t* gpu_elemental_shape_source(ASR::expr_t *e) {
-    if (!e) return nullptr;
-    ASR::expr_t *operands[2] = {nullptr, nullptr};
-    size_t n_operands = 0;
-    switch (e->type) {
-        case ASR::exprType::IntegerBinOp: {
-            ASR::IntegerBinOp_t *o =
-                ASR::down_cast<ASR::IntegerBinOp_t>(e);
-            operands[0] = o->m_left; operands[1] = o->m_right;
-            n_operands = 2; break;
-        }
-        case ASR::exprType::RealBinOp: {
-            ASR::RealBinOp_t *o = ASR::down_cast<ASR::RealBinOp_t>(e);
-            operands[0] = o->m_left; operands[1] = o->m_right;
-            n_operands = 2; break;
-        }
-        case ASR::exprType::ComplexBinOp: {
-            ASR::ComplexBinOp_t *o =
-                ASR::down_cast<ASR::ComplexBinOp_t>(e);
-            operands[0] = o->m_left; operands[1] = o->m_right;
-            n_operands = 2; break;
-        }
-        case ASR::exprType::LogicalBinOp: {
-            ASR::LogicalBinOp_t *o =
-                ASR::down_cast<ASR::LogicalBinOp_t>(e);
-            operands[0] = o->m_left; operands[1] = o->m_right;
-            n_operands = 2; break;
-        }
-        case ASR::exprType::IntegerUnaryMinus: {
-            operands[0] =
-                ASR::down_cast<ASR::IntegerUnaryMinus_t>(e)->m_arg;
-            n_operands = 1; break;
-        }
-        case ASR::exprType::RealUnaryMinus: {
-            operands[0] = ASR::down_cast<ASR::RealUnaryMinus_t>(e)->m_arg;
-            n_operands = 1; break;
-        }
-        case ASR::exprType::ComplexUnaryMinus: {
-            operands[0] =
-                ASR::down_cast<ASR::ComplexUnaryMinus_t>(e)->m_arg;
-            n_operands = 1; break;
-        }
-        case ASR::exprType::LogicalNot: {
-            operands[0] = ASR::down_cast<ASR::LogicalNot_t>(e)->m_arg;
-            n_operands = 1; break;
-        }
-        case ASR::exprType::Cast: {
-            operands[0] = ASR::down_cast<ASR::Cast_t>(e)->m_arg;
-            n_operands = 1; break;
-        }
-        case ASR::exprType::IntrinsicElementalFunction: {
-            ASR::IntrinsicElementalFunction_t *f =
-                ASR::down_cast<ASR::IntrinsicElementalFunction_t>(e);
-            for (size_t i = 0; i < f->n_args; i++) {
-                ASR::expr_t *a =
-                    ASRUtils::get_past_array_physical_cast(f->m_args[i]);
-                if (!a || ASR::is_a<ASR::ArrayBroadcast_t>(*a)) continue;
-                if (ASRUtils::is_array(ASRUtils::expr_type(a))) return a;
-            }
-            return nullptr;
-        }
-        default: return nullptr;
-    }
-    for (size_t i = 0; i < n_operands; i++) {
-        ASR::expr_t *a = ASRUtils::get_past_array_physical_cast(operands[i]);
-        if (!a || ASR::is_a<ASR::ArrayBroadcast_t>(*a)) continue;
-        if (ASRUtils::is_array(ASRUtils::expr_type(a))) return a;
-    }
-    return nullptr;
-}
-
-static ASR::expr_t* find_gpu_first_assigned_value(ASR::stmt_t **body,
-        size_t n_body, const std::string &var_name) {
-    for (size_t i = 0; i < n_body; i++) {
-        if (!ASR::is_a<ASR::Assignment_t>(*body[i])) continue;
-        ASR::Assignment_t *a = ASR::down_cast<ASR::Assignment_t>(body[i]);
-        if (!ASR::is_a<ASR::Var_t>(*a->m_target)) continue;
-        std::string tname = ASRUtils::symbol_name(
-            ASR::down_cast<ASR::Var_t>(a->m_target)->m_v);
-        if (tname != var_name) continue;
-        return a->m_value;
-    }
-    return nullptr;
-}
-
-// The designator whose shape is the shape of the local array `var` of a
-// BLOCK or ASSOCIATE scope, or nullptr when the shape is nowhere to be
-// found.
-//
-// The temporary the frontend materialises for an array-valued ASSOCIATE
-// selector -- `associate(r => sqrt((x-x0)**2 + ...))` -- is an allocatable
-// with deferred extents and no ALLOCATE: the only place its shape is
-// written is the expression assigned to it, and that expression is
-// elementwise, so its shape in turn is that of one of its array operands.
-// Follow the chain down to the designator that actually has extents.
-static ASR::expr_t* gpu_scope_array_shape_source(const ASR::Variable_t *var,
-        ASR::stmt_t **body, size_t n_body) {
-    if (!ASRUtils::is_allocatable(var->m_type)) return nullptr;
-    ASR::ttype_t *inner = ASRUtils::type_get_past_allocatable(var->m_type);
-    if (!ASR::is_a<ASR::Array_t>(*inner)) return nullptr;
-    ASR::Array_t *arr = ASR::down_cast<ASR::Array_t>(inner);
-    if (arr->n_dims == 0) return nullptr;
-    for (size_t d = 0; d < arr->n_dims; d++) {
-        if (arr->m_dims[d].m_length) return nullptr;
-    }
-    std::string vname(var->m_name);
-    if (find_allocate_for_var(body, n_body, vname)) return nullptr;
-    ASR::expr_t *value = find_gpu_first_assigned_value(body, n_body, vname);
-    if (!value) return nullptr;
-    ASR::expr_t *src = ASRUtils::get_past_array_physical_cast(value);
-    while (src) {
-        ASR::expr_t *next = gpu_elemental_shape_source(src);
-        if (!next) break;
-        src = next;
-    }
-    if (!src || !ASRUtils::is_array(ASRUtils::expr_type(src))) {
-        return nullptr;
-    }
-    if (ASRUtils::extract_n_dims_from_ttype(ASRUtils::expr_type(src))
-            != static_cast<int>(arr->n_dims)) {
-        return nullptr;
-    }
-    return src;
-}
-
-
-// Pre-flight for the offload pass: would the backend be able to size every
-// per-thread workspace the loop body needs?
-//
-// The backend describes a workspace with `declared_shape_to_vla_workspace`
-// or `alloc_shape_to_vla_workspace`, and an array whose extents it cannot
-// describe is simply left to the device language, which then reports that
-// the size is not a constant expression -- a code-generation error, far too
-// late, because the offload has been committed to by then. Asking the very
-// same two functions here lets the pass decline instead and leave the loop
-// on the host, and means the pre-flight and the backend cannot disagree.
-static bool gpu_scope_workspaces_resolvable(SymbolTable *symtab,
-        ASR::stmt_t **body, size_t n_body,
-        ASR::stmt_t **root_body, size_t root_n,
-        const std::vector<std::string> &arg_names,
-        std::string &unresolved_name) {
-    if (symtab == nullptr) return true;
-    for (auto &item : symtab->get_scope()) {
-        if (!ASR::is_a<ASR::Variable_t>(*item.second)) continue;
-        ASR::Variable_t *var = ASR::down_cast<ASR::Variable_t>(item.second);
-        ASR::ttype_t *inner = ASRUtils::type_get_past_allocatable(
-            var->m_type);
-        if (!ASR::is_a<ASR::Array_t>(*inner)) continue;
-        ASR::Array_t *arr = ASR::down_cast<ASR::Array_t>(inner);
-        std::string vname(var->m_name);
-        GpuVlaWorkspace ws;
-        ws.var = item.second;
-        if (!ASRUtils::is_allocatable(var->m_type)) {
-            // An automatic array: every extent is in its own type. One
-            // that is a compile-time constant needs no workspace at all.
-            bool runtime = false;
-            for (size_t d = 0; d < arr->n_dims; d++) {
-                if (arr->m_dims[d].m_length &&
-                        !ASR::is_a<ASR::IntegerConstant_t>(
-                            *arr->m_dims[d].m_length)) {
-                    runtime = true;
-                    break;
-                }
-            }
-            if (!runtime) continue;
-            if (declared_shape_to_vla_workspace(arr, vname,
-                    GpuExtentScope{nullptr, arg_names, symtab, root_body,
-                        root_n}, ws)) {
-                continue;
-            }
-            unresolved_name = vname;
-            return false;
-        }
-        ASR::Allocate_t *alloc = find_allocate_for_var(body, n_body, vname);
-        if (alloc == nullptr) continue;
-        ASR::alloc_arg_t *target = find_alloc_arg_for_var(alloc, vname);
-        if (target == nullptr) continue;
-        bool runtime = false;
-        for (size_t d = 0; d < target->n_dims; d++) {
-            if (target->m_dims[d].m_length &&
-                    !ASR::is_a<ASR::IntegerConstant_t>(
-                        *target->m_dims[d].m_length)) {
-                runtime = true;
-                break;
-            }
-        }
-        if (!runtime) continue;
-        if (alloc_shape_to_vla_workspace(*target, arr, vname,
-                GpuExtentScope{nullptr, arg_names, symtab, root_body,
-                    root_n}, ws)) {
-            continue;
-        }
-        unresolved_name = vname;
-        return false;
-    }
-    return true;
-}
-
-// The same question for the whole loop body: the scopes it opens hold the
-// workspaces, so each one is asked in turn.
-static bool gpu_block_workspace_extents_resolvable(
-        ASR::stmt_t **body, size_t n_body,
-        const std::vector<std::string> &arg_names,
-        std::string &unresolved_name) {
-    bool ok = true;
-    gpu_walk_scopes(body, n_body,
-        [&](SymbolTable *symtab, ASR::stmt_t **sbody, size_t sn) {
-            if (!ok) return;
-            if (!gpu_scope_workspaces_resolvable(symtab, sbody, sn,
-                    body, n_body, arg_names, unresolved_name)) {
-                ok = false;
-            }
-        });
-    return ok;
-}
 
 // A new variable named `name` in `scope`, added to it.
 //
@@ -593,254 +266,6 @@ public:
         if (symbols.find(name) == symbols.end()) {
             symbols[name] = {ASRUtils::symbol_type(x.m_v),
                 ASRUtils::EXPR(ASR::make_Var_t(al, x.base.base.loc, x.m_v))};
-        }
-    }
-};
-
-// A derived type is representable only when every one of its data members
-// is, because the device struct is laid out member by member: a single
-// unsupported member anywhere in the type changes the element size the
-// kernel would have to stride by, while the host buffer keeps the wider
-// layout. The members inherited through `extends` live in the parent
-// Struct (they are reached at run time through the `__parent` member), so
-// the parent chain has to be walked as well. `visited` guards against
-// self-referential types such as `type(node), pointer :: next`, whose
-// member graph is cyclic.
-static bool gpu_struct_members_ok(ASR::symbol_t *struct_sym,
-        std::set<ASR::Struct_t*> &visited,
-        const GpuDeviceCapabilities &caps) {
-    ASR::symbol_t *s = ASRUtils::symbol_get_past_external(struct_sym);
-    if (!s || !ASR::is_a<ASR::Struct_t>(*s)) {
-        // The derived type cannot be inspected, so it cannot be shown to
-        // be representable: keep the loop on the CPU.
-        return false;
-    }
-    ASR::Struct_t *st = ASR::down_cast<ASR::Struct_t>(s);
-    if (!visited.insert(st).second) {
-        // Already on the walk stack; its members are checked there.
-        return true;
-    }
-    if (st->m_parent
-            && !gpu_struct_members_ok(st->m_parent, visited, caps)) {
-        return false;
-    }
-    for (size_t i = 0; i < st->n_members; i++) {
-        ASR::symbol_t *msym = st->m_symtab->get_symbol(st->m_members[i]);
-        if (!msym) continue;
-        msym = ASRUtils::symbol_get_past_external(msym);
-        if (!ASR::is_a<ASR::Variable_t>(*msym)) continue;
-        ASR::Variable_t *mvar = ASR::down_cast<ASR::Variable_t>(msym);
-        ASR::ttype_t *mtype = ASRUtils::extract_type(mvar->m_type);
-        if (ASR::is_a<ASR::StructType_t>(*mtype)) {
-            if (!mvar->m_type_declaration
-                    || !gpu_struct_members_ok(
-                        mvar->m_type_declaration, visited, caps)) {
-                return false;
-            }
-        } else if (!caps.has_scalar_type(mtype)) {
-            return false;
-        }
-    }
-    return true;
-}
-
-// Answers whether the selected device can represent the type of `e` with
-// the same in-memory width the host uses. A device whose type set is
-// narrower than the host's -- no 64-bit floating point type, no 64-bit
-// boolean, no complex type -- lowers such data to a narrower (or bogus)
-// type. Offloading a `do concurrent` that touches it would make the kernel
-// reinterpret the host buffers and the by-value scalar-argument struct at
-// the wrong element size, silently producing wrong results, so such a loop
-// has to stay on the CPU.
-//
-// The rule itself is the capability descriptor's, so that the decline this
-// raises and the class that decline is given cannot disagree about what the
-// device has a type for.
-static bool gpu_device_can_represent_type(const GpuDeviceCapabilities &caps,
-        ASR::ttype_t *t, ASR::expr_t *e) {
-    ASR::ttype_t *base_t = ASRUtils::extract_type(t);
-    if (ASR::is_a<ASR::StructType_t>(*base_t)) {
-        if (!e) return false;
-        std::set<ASR::Struct_t*> visited;
-        return gpu_struct_members_ok(
-            ASRUtils::get_struct_sym_from_struct_expr(e), visited, caps);
-    }
-    return caps.has_scalar_type(base_t);
-}
-
-// The scalar element type behind `t`, for a decline that has to be
-// classified against what the device has a type for. A derived type has no
-// single element type -- the width that offends is one member's -- so it
-// answers with nothing, and the decline is classified on its reason alone.
-static ASR::ttype_t* scalar_type_of(ASR::ttype_t *t) {
-    if (!t) return nullptr;
-    ASR::ttype_t *base_t = ASRUtils::extract_type(t);
-    if (ASR::is_a<ASR::StructType_t>(*base_t)) return nullptr;
-    return base_t;
-}
-
-// A variable declared inside the `do concurrent` body by a BLOCK or an
-// ASSOCIATE construct is carried into the generated kernel as a
-// kernel-local declaration. The Metal Shading Language has no
-// variable-length arrays and no heap, so a local whose extent is not a
-// compile-time constant can only be emitted when the GPU workspace
-// machinery (`analyze_gpu_vla_workspaces`) can bind it to a per-thread
-// slice of a device buffer. That needs an extent expression: either the
-// array declares one itself (an automatic array `real :: t(m)`) or an
-// ALLOCATE in the same construct supplies one.
-//
-// What is left over has no extent anywhere: a descriptor local -- the
-// temporary the frontend materialises for an array-valued ASSOCIATE
-// selector whose operand is an allocatable or an assumed-shape dummy, or
-// an allocatable never allocated in this construct -- carries deferred
-// `dimension_t` with null lengths, so the kernel would declare it with a
-// bogus one-element extent and index out of bounds. Such a loop has to
-// stay on the CPU.
-class GpuLocalArrayChecker :
-        public ASR::BaseWalkVisitor<GpuLocalArrayChecker> {
-public:
-    bool has_unsized_local_array = false;
-    std::string unsized_name;
-
-    // True when every extent of the array is present as an expression,
-    // so a workspace buffer can be sized from it.
-    static bool dims_have_lengths(ASR::dimension_t *dims, size_t n) {
-        if (n == 0) return false;
-        for (size_t d = 0; d < n; d++) {
-            if (!dims[d].m_length) return false;
-        }
-        return true;
-    }
-
-    void check_scope(SymbolTable *symtab, ASR::stmt_t **body,
-            size_t n_body) {
-        if (!symtab) return;
-        for (auto &item : symtab->get_scope()) {
-            ASR::symbol_t *sym = item.second;
-            if (!sym || !ASR::is_a<ASR::Variable_t>(*sym)) continue;
-            ASR::Variable_t *var = ASR::down_cast<ASR::Variable_t>(sym);
-            ASR::ttype_t *t =
-                ASRUtils::type_get_past_allocatable_pointer(var->m_type);
-            ASR::dimension_t *dims = nullptr;
-            int rank = ASRUtils::extract_dimensions_from_ttype(t, dims);
-            if (rank <= 0) continue;
-            if (ASRUtils::is_fixed_size_array(dims, rank)) continue;
-            if (dims_have_lengths(dims, (size_t)rank)) continue;
-            if (ASRUtils::is_allocatable(var->m_type)) {
-                ASR::Allocate_t *alloc = find_allocate_for_var(body,
-                    n_body, std::string(var->m_name));
-                ASR::alloc_arg_t *aa = alloc
-                    ? find_alloc_arg_for_var(alloc,
-                        std::string(var->m_name))
-                    : nullptr;
-                if (aa && dims_have_lengths(aa->m_dims, aa->n_dims)) {
-                    continue;
-                }
-                // No shape anywhere, but the expression assigned to it
-                // has one: size_scope_array_temporaries() writes that
-                // shape into the temporary's own type before the kernel
-                // is built, and the workspace pre-flight then decides
-                // whether the host can evaluate it.
-                if (gpu_scope_array_shape_source(var, body, n_body)) {
-                    continue;
-                }
-            }
-            has_unsized_local_array = true;
-            if (unsized_name.empty()) unsized_name = item.first;
-        }
-    }
-
-    void visit_BlockCall(const ASR::BlockCall_t &x) {
-        ASR::symbol_t *b = ASRUtils::symbol_get_past_external(x.m_m);
-        if (!b || !ASR::is_a<ASR::Block_t>(*b)) return;
-        ASR::Block_t *blk = ASR::down_cast<ASR::Block_t>(b);
-        check_scope(blk->m_symtab, blk->m_body, blk->n_body);
-        for (size_t i = 0; i < blk->n_body; i++) {
-            visit_stmt(*blk->m_body[i]);
-        }
-    }
-
-    void visit_AssociateBlockCall(const ASR::AssociateBlockCall_t &x) {
-        ASR::symbol_t *b = ASRUtils::symbol_get_past_external(x.m_m);
-        if (!b || !ASR::is_a<ASR::AssociateBlock_t>(*b)) return;
-        ASR::AssociateBlock_t *blk =
-            ASR::down_cast<ASR::AssociateBlock_t>(b);
-        check_scope(blk->m_symtab, blk->m_body, blk->n_body);
-        for (size_t i = 0; i < blk->n_body; i++) {
-            visit_stmt(*blk->m_body[i]);
-        }
-    }
-};
-
-// A BLOCK-local is not a kernel argument, so GpuSymbolCollector skips it
-// and the representability sweep over kernel-reaching symbols never sees
-// it. The device's type set still has to apply: a real(16) local used to
-// compile as float at a 16-byte host stride, and a derived-type local whose
-// member is real(8) or complex is the same hole.
-class GpuLocalWidthChecker :
-        public ASR::BaseWalkVisitor<GpuLocalWidthChecker> {
-public:
-    bool unsupported = false;
-    std::string bad_name;
-    // The element type that was turned down, when it is a scalar one. A
-    // derived type leaves this null: the width that offends is a member's,
-    // and the message names the local rather than a type.
-    ASR::ttype_t *bad_type = nullptr;
-    // What the selected device has a scalar type of.
-    GpuDeviceCapabilities caps;
-
-    bool type_ok(ASR::Variable_t *var, ASR::ttype_t *&offending) {
-        ASR::ttype_t *base = ASRUtils::extract_type(var->m_type);
-        if (ASR::is_a<ASR::StructType_t>(*base)) {
-            // A BLOCK-local is not a kernel argument, so the symbol
-            // collector never hands this type to
-            // gpu_device_can_represent_type. Walk the members here: a
-            // real(8) or complex component is the same silent-wrong-width
-            // hole the scalar path already closed.
-            if (!var->m_type_declaration) return false;
-            std::set<ASR::Struct_t*> visited;
-            return gpu_struct_members_ok(var->m_type_declaration, visited,
-                caps);
-        }
-        offending = base;
-        return caps.has_scalar_type(base);
-    }
-
-    void check_scope(SymbolTable *symtab) {
-        if (!symtab) return;
-        for (auto &item : symtab->get_scope()) {
-            if (!ASR::is_a<ASR::Variable_t>(*item.second)) continue;
-            ASR::Variable_t *var = ASR::down_cast<ASR::Variable_t>(
-                item.second);
-            ASR::ttype_t *offending = nullptr;
-            if (!type_ok(var, offending)) {
-                unsupported = true;
-                if (bad_name.empty()) {
-                    bad_name = item.first;
-                    bad_type = offending;
-                }
-            }
-        }
-    }
-
-    void visit_BlockCall(const ASR::BlockCall_t &x) {
-        ASR::symbol_t *b = ASRUtils::symbol_get_past_external(x.m_m);
-        if (!b || !ASR::is_a<ASR::Block_t>(*b)) return;
-        ASR::Block_t *blk = ASR::down_cast<ASR::Block_t>(b);
-        check_scope(blk->m_symtab);
-        for (size_t i = 0; i < blk->n_body; i++) {
-            visit_stmt(*blk->m_body[i]);
-        }
-    }
-
-    void visit_AssociateBlockCall(const ASR::AssociateBlockCall_t &x) {
-        ASR::symbol_t *b = ASRUtils::symbol_get_past_external(x.m_m);
-        if (!b || !ASR::is_a<ASR::AssociateBlock_t>(*b)) return;
-        ASR::AssociateBlock_t *blk =
-            ASR::down_cast<ASR::AssociateBlock_t>(b);
-        check_scope(blk->m_symtab);
-        for (size_t i = 0; i < blk->n_body; i++) {
-            visit_stmt(*blk->m_body[i]);
         }
     }
 };
@@ -1506,51 +931,6 @@ public:
     }
 };
 
-// A statement the device has no way to run. A kernel that held one would
-// simply not run it, so the effect the program asked for would go missing
-// with nothing to show for it.
-static GpuDeclineReason unsupported_on_device(const ASR::stmt_t &s) {
-    switch (s.type) {
-        case ASR::stmtType::Print:
-        case ASR::stmtType::FileWrite:
-        case ASR::stmtType::FileRead:
-        case ASR::stmtType::FileOpen:
-        case ASR::stmtType::FileClose:
-        case ASR::stmtType::FileInquire:
-        case ASR::stmtType::FileBackspace:
-        case ASR::stmtType::FileRewind:
-        case ASR::stmtType::FileEndfile:
-        case ASR::stmtType::Flush:
-            return GpuDeclineReason::StatementIo;
-        case ASR::stmtType::Stop:
-        case ASR::stmtType::ErrorStop:
-            return GpuDeclineReason::StatementStop;
-        default:
-            return GpuDeclineReason::None;
-    }
-}
-
-// The first statement of a body that the device cannot run, and where it is.
-class GpuUnsupportedStatementFinder
-        : public ASRUtils::BlockBodyWalkVisitor<GpuUnsupportedStatementFinder> {
-public:
-    GpuDeclineReason reason;
-    Location loc;
-
-    GpuUnsupportedStatementFinder() : reason(GpuDeclineReason::None) {}
-
-    void visit_stmt(const ASR::stmt_t &s) {
-        if (reason != GpuDeclineReason::None) return;
-        GpuDeclineReason why = unsupported_on_device(s);
-        if (why != GpuDeclineReason::None) {
-            reason = why;
-            loc = s.base.loc;
-            return;
-        }
-        ASR::BaseWalkVisitor<GpuUnsupportedStatementFinder>::visit_stmt(s);
-    }
-};
-
 // Every routine a loop body reaches, however deep.
 static std::vector<ASR::Function_t*> reachable_routines(ASR::stmt_t **body,
         size_t n_body) {
@@ -1600,155 +980,6 @@ public:
         for (ASR::OMPRegion_t *region : loops) {
             region->m_exec_target = ASR::exec_targetType::ExecSerial;
         }
-    }
-};
-
-// Metal shaders have neither variable-length arrays nor a heap, so a
-// device function (an `inline` callee of a kernel) can only declare
-// locals whose extent the shader compiler can fold to a constant. An
-// array constructor whose elements are themselves array-valued
-// expressions is materialized by the later `array_struct_temporary`
-// pass into a temporary sized from those elements. An element sized
-// from an assumed-shape or deferred-shape dummy argument, or from a
-// local allocatable whose ALLOCATE bounds are themselves only known at
-// run time, would have to be a VLA inside the device function -- which
-// Metal cannot express. Detect that shape here so the loop can be
-// declined and run on the host instead. Elements sized from a local
-// allocatable with constant ALLOCATE bounds are fine: the Metal backend
-// resolves those extents from the ALLOCATE statement.
-class GpuDeviceFunctionArrayTempChecker :
-        public ASRUtils::BlockBodyWalkVisitor<GpuDeviceFunctionArrayTempChecker> {
-private:
-
-    // Entities of the function being checked whose extent is not a
-    // compile-time constant: assumed-shape/deferred-shape dummy
-    // arguments, and local allocatables whose ALLOCATE bounds are only
-    // known at run time.
-    std::set<ASR::symbol_t*> runtime_extent_syms;
-
-    static bool has_runtime_extent(ASR::ttype_t *t) {
-        if (!t) return false;
-        ASR::dimension_t *dims = nullptr;
-        int rank = ASRUtils::extract_dimensions_from_ttype(
-            ASRUtils::type_get_past_allocatable_pointer(t), dims);
-        return rank > 0 && !ASRUtils::is_fixed_size_array(dims, rank);
-    }
-
-    class VarRefCollector :
-            public ASR::BaseWalkVisitor<VarRefCollector> {
-    public:
-        std::set<ASR::symbol_t*> vars;
-
-        void visit_Var(const ASR::Var_t &x) {
-            vars.insert(ASRUtils::symbol_get_past_external(x.m_v));
-        }
-    };
-
-    // Collects the targets of ALLOCATE statements whose bounds are not
-    // compile-time constants, descending into BLOCK and ASSOCIATE
-    // scopes so an allocation nested there is seen too.
-    class RuntimeAllocCollector :
-            public ASRUtils::BlockBodyWalkVisitor<RuntimeAllocCollector> {
-    public:
-        std::set<ASR::symbol_t*> vars;
-
-        void visit_Allocate(const ASR::Allocate_t &x) {
-            for (size_t i = 0; i < x.n_args; i++) {
-                if (!x.m_args[i].m_a ||
-                        !ASR::is_a<ASR::Var_t>(*x.m_args[i].m_a)) continue;
-                bool all_const = true;
-                for (size_t d = 0; d < x.m_args[i].n_dims; d++) {
-                    ASR::expr_t *len = x.m_args[i].m_dims[d].m_length;
-                    if (!len || !ASRUtils::expr_value(len)) {
-                        all_const = false;
-                    }
-                }
-                if (all_const) continue;
-                vars.insert(ASRUtils::symbol_get_past_external(
-                    ASR::down_cast<ASR::Var_t>(x.m_args[i].m_a)->m_v));
-            }
-        }
-
-    };
-
-public:
-    bool has_runtime_sized_temp = false;
-
-    void check_function(ASR::Function_t *fn) {
-        runtime_extent_syms.clear();
-        for (size_t i = 0; i < fn->n_args; i++) {
-            if (!ASR::is_a<ASR::Var_t>(*fn->m_args[i])) continue;
-            ASR::symbol_t *sym = ASRUtils::symbol_get_past_external(
-                ASR::down_cast<ASR::Var_t>(fn->m_args[i])->m_v);
-            if (!sym || !ASR::is_a<ASR::Variable_t>(*sym)) continue;
-            ASR::Variable_t *var = ASR::down_cast<ASR::Variable_t>(sym);
-            ASR::ttype_t *bare = ASRUtils::extract_type(var->m_type);
-            // A struct dummy carries the extents of its allocatable
-            // components in run-time parameters too, so an element sized
-            // from one of its components is just as unrepresentable.
-            if (has_runtime_extent(var->m_type) ||
-                    ASR::is_a<ASR::StructType_t>(*bare) ||
-                    ASRUtils::is_class_type(bare)) {
-                runtime_extent_syms.insert(sym);
-            }
-        }
-        std::set<ASR::symbol_t*> passed_in;
-        passed_in.insert(runtime_extent_syms.begin(),
-            runtime_extent_syms.end());
-        // The result is not a dummy yet. subroutine_from_function later
-        // turns it into one and plants a VLA temporary at the call site
-        // in the kernel. That temporary has to be spliced into the
-        // kernel (or the loop declined) so pre-flight can size it.
-        {
-            RuntimeAllocCollector alloc_collector;
-            for (size_t i = 0; i < fn->n_body; i++) {
-                alloc_collector.visit_stmt(*fn->m_body[i]);
-            }
-            // A local array the callee sizes at run time is a
-            // variable-length array in the shader, and Metal has no
-            // declaration for one. This is how an array constructor
-            // already lowered to a temporary reaches here: the
-            // constructor is gone, and only the temporary is left.
-            // Splicing the callee into the kernel moves the temporary
-            // to a scope the per-thread workspace machinery can size.
-            for (ASR::symbol_t *v : alloc_collector.vars) {
-                if (!passed_in.count(v)) has_runtime_sized_temp = true;
-            }
-            runtime_extent_syms.insert(alloc_collector.vars.begin(),
-                alloc_collector.vars.end());
-        }
-        for (auto &item : fn->m_symtab->get_scope()) {
-            ASR::symbol_t *sym = item.second;
-            if (!ASR::is_a<ASR::Variable_t>(*sym)) continue;
-            if (passed_in.count(sym)) continue;
-            ASR::Variable_t *var = ASR::down_cast<ASR::Variable_t>(sym);
-            if (var->m_intent != ASR::intentType::Local) continue;
-            if (ASRUtils::is_allocatable(var->m_type)) continue;
-            if (has_runtime_extent(var->m_type)) {
-                has_runtime_sized_temp = true;
-            }
-        }
-        if (runtime_extent_syms.empty()) return;
-        for (size_t i = 0; i < fn->n_body; i++) {
-            visit_stmt(*fn->m_body[i]);
-        }
-    }
-
-    void visit_ArrayConstructor(const ASR::ArrayConstructor_t &x) {
-        for (size_t i = 0; i < x.n_args; i++) {
-            ASR::expr_t *arg = x.m_args[i];
-            if (!arg) continue;
-            if (!has_runtime_extent(ASRUtils::expr_type(arg))) continue;
-            VarRefCollector vc;
-            vc.visit_expr(*arg);
-            for (ASR::symbol_t *v : vc.vars) {
-                if (runtime_extent_syms.count(v)) {
-                    has_runtime_sized_temp = true;
-                }
-            }
-        }
-        ASR::BaseWalkVisitor<GpuDeviceFunctionArrayTempChecker>::
-            visit_ArrayConstructor(x);
     }
 };
